+++ lib/layer/layer.dart
@@ -1,5 +1,8 @@
-import 'dart:ui';
 import 'package:meta/meta.dart';
+import 'dart:ui';
+
+import './processors.dart';
+export './processors.dart';
 
 abstract class Layer {
   List<LayerProcessor> preProcessors = [];
@@ -26,8 +29,10 @@ abstract class Layer {
   }
 
   Canvas get canvas {
-    assert(_canvas != null,
-        'Layer is not ready for rendering, call beginRendering first');
+    assert(
+      _canvas != null,
+      'Layer is not ready for rendering, call beginRendering first',
+    );
     return _canvas;
   }
 
@@ -62,33 +67,5 @@ abstract class DynamicLayer extends Layer {
     finishRendering();
 
     super.render(canvas, x: x, y: y);
-  }
-}
-
-abstract class LayerProcessor {
-  void process(Picture pic, Canvas canvas);
-}
-
-class ShadowProcessor extends LayerProcessor {
-  Paint _shadowPaint;
-
-  final Offset offset;
-
-  ShadowProcessor({
-    this.offset = const Offset(10, 10),
-    double opacity = 0.9,
-    Color color = const Color(0xFF000000),
-  }) {
-    _shadowPaint = Paint()
-      ..colorFilter =
-          ColorFilter.mode(color.withOpacity(opacity), BlendMode.srcATop);
-  }
-
-  @override
-  void process(Picture pic, Canvas canvas) {
-    canvas.saveLayer(Rect.largest, _shadowPaint);
-    canvas.translate(offset.dx, offset.dy);
-    canvas.drawPicture(pic);
-    canvas.restore();
   }
 }
